refactor: split utils in multiple files
